feat(rollup): add assets detection on "src" attributes
Create the "glade-core" taglib with it's first transformer to detect :
    <tag src=".LOCAL_PATH" /> (catching both "./PATH" & "../PATH")
The standard Rollup URL plugin deals with the files and the GladeJS plugin
collects the paths mapping code to apply it on pages and components code.

Reorganised the "pages" to demonstrate the usage with Marko logos
(the PNG is copied/renamed, the SVG is inlined being less than 10 kb).
Also, quick version bumps to marko, rollup, glob & semantic-release.